Rename functions to executor, correct log for duration from int to string
